pyproject.toml
# pyproject.toml
# Packaging metadata for 21_home_temp. uv relies on this file to lock deps.
[project]
name = "21-home-temp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "libsql-experimental>=0.0.55",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
